requirements.txt
certifi==2024.6.2
charset-normalizer==3.3.2
idna==3.7
jmespath==1.0.1
load-dotenv==0.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
requests==2.32.3
six==1.16.0
urllib3==2.2.1
google-cloud-storage